+++ internal/config/config_test.go
@@ -1,12 +1,10 @@
 package config
 
 import (
-	"fmt"
 	"github.com/stretchr/testify/assert"
 	"os"
 	"path"
 	"testing"
-	"time"
 )
 
 var exampleDir string
@@ -45,52 +43,4 @@ func TestNewConfigsFromDir(t *testing.T) {
 		WithJustValidation(true),
 	)
 	assert.NoError(t, err)
-}
-
-func TestSchedule_SetDefaults_Schedule(t *testing.T) {
-	conf := &Config{}
-	defaults(conf)
-	assert.Equal(t, TypeSchedulerStrategyTick, conf.Schedule.Strategy)
-}
-
-func TestSchedule_Validate(t *testing.T) {
-	h := time.Hour
-	c := "* * * * *"
-
-	testCases := []struct {
-		name     string
-		schedule Schedule
-		err      error
-	}{
-		{
-			name:     "must_not_miss_interval_and_cron",
-			schedule: Schedule{},
-			err:      fmt.Errorf("must set schedule.interval or schedule.cron"),
-		},
-		{
-			name: "must_not_have_interval_and_cron",
-			schedule: Schedule{
-				Interval: &h,
-				Cron:     &c,
-			},
-			err: fmt.Errorf("must set either schedule.interval or schedule.cro"),
-		},
-		{
-			name: "must_have_valid_strategy",
-			schedule: Schedule{
-				Cron: &c,
-			},
-			err: fmt.Errorf("unknown strategy: \"\""),
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			err := tc.schedule.Validate()
-			if tc.err == nil {
-				assert.NoError(t, err)
-			} else {
-				assert.Error(t, tc.err, err)
-			}
-		})
-	}
 }
